Add test refactoring skeleton
