+++ lib/cf/renewer/src/index.js
@@ -265,41 +265,28 @@ const processUsage = (reference, systemToken, failure, cb) => {
       return;
     }
 
-    const rawStopUsage = zeroUsage(refreshUsage(doc, 0), 'current');
-    const stopUsage = sanitizeUsageDoc(rawStopUsage);
-    debug('Reporting stop usage for doc with id %s ...', doc.id);
-    reportUsage(stopUsage, systemToken, (error, response) => {
+    const rawStartUsage = zeroUsage(refreshUsage(doc, 0), 'previous');
+    const startUsage = sanitizeUsageDoc(rawStartUsage);
+    debug('Reporting start usage for doc with id %s ...', doc.id);
+    reportUsage(startUsage, systemToken, (error, response) => {
       if (error) {
-        failure(buildError('stop report', stopUsage, error, response),
+        failure(buildError('start report', startUsage, error, response),
           response);
         return;
       }
-      if (response.statusCode == 409)
-        debug('Skipped conflicting stop usage with id %s', doc.id);
+      if (response.statusCode == 409) {
+        debug('Skipped conflicting start usage with id %s', doc.id);
+        cb();
+        return;
+      }
 
-      const rawStartUsage = zeroUsage(refreshUsage(doc, 1), 'previous');
-      const startUsage = sanitizeUsageDoc(rawStartUsage);
-      debug('Reporting start usage for doc with id %s ...', doc.id);
-      reportUsage(startUsage, systemToken, (error, response) => {
+      carryOver.write(startUsage, response, (error) => {
         if (error) {
-          failure(buildError('start report', startUsage, error, response),
-            response);
-          return;
-        }
-        if (response.statusCode == 409) {
-          debug('Skipped conflicting start usage with id %s', doc.id);
-          cb();
+          failure(error, response);
           return;
         }
 
-        carryOver.write(startUsage, response, (error) => {
-          if (error) {
-            failure(error, response);
-            return;
-          }
-
-          cb();
-        });
+        cb();
       });
     });
   });
