+++ lib/cf/renewer/src/test/report-usage-test.js
@@ -210,17 +210,15 @@ const tests = (secured) => {
 
       it('reports refreshed usage to COLLECTOR', () => {
         const args = reqmock.post.args;
-        expect(args.length).to.equal(4);
-        checkPostRequest(args[0], generateUsage(appUsage, 0, 'current'));
-        checkPostRequest(args[1], generateUsage(appUsage, 1, 'previous'));
-        checkPostRequest(args[2], generateUsage(appUsage, 0, 'current'));
-        checkPostRequest(args[3], generateUsage(appUsage, 1, 'previous'));
+        expect(args.length).to.equal(2);
+        checkPostRequest(args[0], generateUsage(appUsage, 0, 'previous'));
+        checkPostRequest(args[1], generateUsage(appUsage, 0, 'previous'));
       });
 
       it('counts the reported usage', () => {
         expect(renewer.statistics.usage.getSuccess).to.equal(2);
         expect(renewer.statistics.usage.getFailures).to.equal(0);
-        expect(renewer.statistics.usage.reportSuccess).to.equal(4);
+        expect(renewer.statistics.usage.reportSuccess).to.equal(2);
         expect(renewer.statistics.usage.reportConflict).to.equal(0);
         expect(renewer.statistics.usage.reportFailures).to.equal(0);
       });
@@ -357,15 +355,14 @@ const tests = (secured) => {
 
         it('reports usage to COLLECTOR', () => {
           const args = reqmock.post.args;
-          expect(args.length).to.equal(2);
-          checkPostRequest(args[0], generateUsage(appUsage, 0, 'current'));
-          checkPostRequest(args[1], generateUsage(appUsage, 1, 'previous'));
+          expect(args.length).to.equal(1);
+          checkPostRequest(args[0], generateUsage(appUsage, 0, 'previous'));
         });
 
         it('counts the outcome', () => {
           expect(renewer.statistics.usage.getSuccess).to.equal(1);
           expect(renewer.statistics.usage.getFailures).to.equal(1);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(2);
+          expect(renewer.statistics.usage.reportSuccess).to.equal(1);
           expect(renewer.statistics.usage.reportConflict).to.equal(0);
           expect(renewer.statistics.usage.reportFailures).to.equal(0);
         });
@@ -563,15 +560,14 @@ const tests = (secured) => {
 
         it('reports refreshed usage to COLLECTOR', () => {
           const args = reqmock.post.args;
-          expect(args.length).to.equal(2);
-          checkPostRequest(args[0], generateUsage(appUsage, 0, 'current'));
-          checkPostRequest(args[1], generateUsage(appUsage, 1, 'previous'));
+          expect(args.length).to.equal(1);
+          checkPostRequest(args[0], generateUsage(appUsage, 0, 'previous'));
         });
 
         it('counts the outcome', () => {
           expect(renewer.statistics.usage.getSuccess).to.equal(1);
           expect(renewer.statistics.usage.getFailures).to.equal(1);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(2);
+          expect(renewer.statistics.usage.reportSuccess).to.equal(1);
           expect(renewer.statistics.usage.reportConflict).to.equal(0);
           expect(renewer.statistics.usage.reportFailures).to.equal(0);
         });
@@ -761,81 +757,13 @@ const tests = (secured) => {
 
         it('reports refreshed usage to COLLECTOR', () => {
           const args = reqmock.post.args;
-          expect(args.length).to.equal(3);
-
+          expect(args.length).to.equal(1);
           checkPostRequest(args[0],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '1'));
-          checkPostRequest(args[1],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '1'));
-          checkPostRequest(args[2],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '2'));
+            changeOrgId(generateUsage(appUsage, 0, 'previous'), '1'));
         });
 
         it('counts the outcome', () => {
           expect(renewer.statistics.usage.getSuccess).to.equal(2);
-          expect(renewer.statistics.usage.getFailures).to.equal(0);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(2);
-          expect(renewer.statistics.usage.reportConflict).to.equal(0);
-          expect(renewer.statistics.usage.reportFailures).to.equal(1);
-        });
-
-        it('raises correct reporting errors', () => {
-          expect(renewer.errors.noReportEverHappened).to.equal(false);
-          expect(renewer.errors.noGetEverHappened).to.equal(false);
-
-          expect(renewer.errors.lastError).to.match(/^Failed reporting usage/);
-          const errorTime = moment(renewer.errors.lastErrorTimestamp)
-            .valueOf();
-          expect(errorTime).to.be.at.least(startTime);
-
-          expect(renewer.errors.consecutiveGetFailures).to.equal(0);
-          expect(renewer.errors.consecutiveReportFailures).to.equal(1);
-        });
-      });
-
-      context('on the second usage post (start)', () => {
-        const startTime = moment.now();
-
-        beforeEach((done) => {
-          numPostRequestToError = 2;
-
-          renewer = require('..');
-          renewer.renewUsage(systemToken, {
-            failure: (error, response) => {
-              renewer.stopRenewer();
-
-              expect(error.op).to.equal('start report');
-              expect(error.doc).to.not.equal(undefined);
-              expect(error.error).to.equal('error');
-              expect(error.response).to.deep.equal(mockedResponse);
-              expect(response).to.deep.equal(mockedResponse);
-              done();
-            },
-            success: () => {
-              renewer.stopRenewer();
-              done(new Error('Unexpected call of success'));
-            }
-          });
-        });
-
-        it('gets the real usage from the COLLECTOR', () => {
-          const args = reqmock.get.args;
-          expect(args.length).to.equal(1);
-          checkGetRequest(args[0], '1');
-        });
-
-        it('reports refreshed usage to COLLECTOR', () => {
-          const args = reqmock.post.args;
-          expect(args.length).to.equal(2);
-
-          checkPostRequest(args[0],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '1'));
-          checkPostRequest(args[1],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '1'));
-        });
-
-        it('counts the outcome', () => {
-          expect(renewer.statistics.usage.getSuccess).to.equal(1);
           expect(renewer.statistics.usage.getFailures).to.equal(0);
           expect(renewer.statistics.usage.reportSuccess).to.equal(1);
           expect(renewer.statistics.usage.reportConflict).to.equal(0);
@@ -1040,19 +968,15 @@ const tests = (secured) => {
 
         it('reports refreshed usage to COLLECTOR', () => {
           const args = reqmock.post.args;
-          expect(args.length).to.equal(3);
+          expect(args.length).to.equal(1);
           checkPostRequest(args[0],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '1'));
-          checkPostRequest(args[1],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '1'));
-          checkPostRequest(args[2],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '2'));
+            changeOrgId(generateUsage(appUsage, 0, 'previous'), '1'));
         });
 
         it('counts the outcome', () => {
           expect(renewer.statistics.usage.getSuccess).to.equal(2);
           expect(renewer.statistics.usage.getFailures).to.equal(0);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(2);
+          expect(renewer.statistics.usage.reportSuccess).to.equal(1);
           expect(renewer.statistics.usage.reportConflict).to.equal(0);
           expect(renewer.statistics.usage.reportFailures).to.equal(1);
         });
@@ -1073,7 +997,7 @@ const tests = (secured) => {
 
       context('when 409 is returned', () => {
         beforeEach((done) => {
-          numPostRequestToError = 3;
+          numPostRequestToError = 2;
           errorResponseCode = 409;
 
           renewer = require('..');
@@ -1099,21 +1023,17 @@ const tests = (secured) => {
 
         it('reports refreshed usage to COLLECTOR', () => {
           const args = reqmock.post.args;
-          expect(args.length).to.equal(4);
+          expect(args.length).to.equal(2);
           checkPostRequest(args[0],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '1'));
+            changeOrgId(generateUsage(appUsage, 0, 'previous'), '1'));
           checkPostRequest(args[1],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '1'));
-          checkPostRequest(args[2],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '2'));
-          checkPostRequest(args[3],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '2'));
+            changeOrgId(generateUsage(appUsage, 0, 'previous'), '2'));
         });
 
         it('counts the outcome', () => {
           expect(renewer.statistics.usage.getSuccess).to.equal(2);
           expect(renewer.statistics.usage.getFailures).to.equal(0);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(3);
+          expect(renewer.statistics.usage.reportSuccess).to.equal(1);
           expect(renewer.statistics.usage.reportConflict).to.equal(1);
           expect(renewer.statistics.usage.reportFailures).to.equal(0);
         });
@@ -1124,71 +1044,6 @@ const tests = (secured) => {
 
           expect(renewer.errors.consecutiveGetFailures).to.equal(0);
           expect(renewer.errors.consecutiveReportFailures).to.equal(0);
-        });
-      });
-
-      context('on the second usage post (start)', () => {
-        beforeEach((done) => {
-          numPostRequestToError = 2;
-          errorResponseCode = 500;
-
-          renewer = require('..');
-          renewer.renewUsage(systemToken, {
-            failure: (error, response) => {
-              renewer.stopRenewer();
-
-              expect(error.op).to.equal('start report');
-              expect(error.doc).not.to.equal(undefined);
-              expect(error.error).not.to.equal(undefined);
-              expect(error.response).to.deep.equal(
-                getResponse(errorResponseCode)
-              );
-              expect(response).to.deep.equal(
-                getResponse(errorResponseCode)
-              );
-              done();
-            },
-            success: () => {
-              renewer.stopRenewer();
-              done(new Error('Unexpected call of success'));
-            }
-          });
-        });
-
-        it('gets the real usage from the COLLECTOR', () => {
-          const args = reqmock.get.args;
-          expect(args.length).to.equal(1);
-          checkGetRequest(args[0], '1');
-        });
-
-        it('reports refreshed usage to COLLECTOR', () => {
-          const args = reqmock.post.args;
-          expect(args.length).to.equal(2);
-          checkPostRequest(args[0],
-            changeOrgId(generateUsage(appUsage, 0, 'current'), '1'));
-          checkPostRequest(args[1],
-            changeOrgId(generateUsage(appUsage, 1, 'previous'), '1'));
-        });
-
-        it('counts the outcome', () => {
-          expect(renewer.statistics.usage.getSuccess).to.equal(1);
-          expect(renewer.statistics.usage.getFailures).to.equal(0);
-          expect(renewer.statistics.usage.reportSuccess).to.equal(1);
-          expect(renewer.statistics.usage.reportConflict).to.equal(0);
-          expect(renewer.statistics.usage.reportFailures).to.equal(1);
-        });
-
-        it('raises correct reporting errors', () => {
-          expect(renewer.errors.noGetEverHappened).to.equal(false);
-          expect(renewer.errors.noReportEverHappened).to.equal(false);
-
-          expect(renewer.errors.lastError).to.match(/^Failed reporting usage/);
-          const errorTime = moment(renewer.errors.lastErrorTimestamp)
-            .valueOf();
-          expect(errorTime).to.be.at.least(startTime);
-
-          expect(renewer.errors.consecutiveGetFailures).to.equal(0);
-          expect(renewer.errors.consecutiveReportFailures).to.equal(1);
         });
       });
