+++ lib/cf/renewer/src/test/retry-usage-test.js
@@ -174,7 +174,7 @@ const tests = (secured) => {
         failure: (error, response) => {
           renewer.stopRenewer();
 
-          expect(error.op).to.equal('stop report');
+          expect(error.op).to.equal('start report');
           expect(error.doc).to.not.equal(undefined);
           expect(error.error).to.equal('error');
           expect(error.response).to.deep.equal(okPostResponse);
@@ -198,7 +198,7 @@ const tests = (secured) => {
         missingToken: 0,
         getSuccess: 3,
         getFailures: 0,
-        reportSuccess: 4,
+        reportSuccess: 2,
         reportConflict: 0,
         reportFailures: 1
       });
@@ -223,7 +223,7 @@ const tests = (secured) => {
         failure: (error, response) => {
           renewer.stopRenewer();
 
-          expect(error.op).to.equal('stop report');
+          expect(error.op).to.equal('start report');
           expect(error.doc).to.not.equal(undefined);
           expect(error.error).to.equal('error');
           expect(error.response).to.deep.equal(okPostResponse);
@@ -247,7 +247,7 @@ const tests = (secured) => {
         missingToken: 0,
         getSuccess: 2,
         getFailures: 0,
-        reportSuccess: 2,
+        reportSuccess: 1,
         reportConflict: 0,
         reportFailures: 1
       });
@@ -313,7 +313,7 @@ const tests = (secured) => {
               missingToken: 0,
               getFailures: 0,
               getSuccess: 6,
-              reportSuccess: 8,
+              reportSuccess: 4,
               reportConflict: 0,
               reportFailures: 2
             });
