--- test/cf/bridge/README.md
@@ -1,4 +1,0 @@
-abacus-cf-bridge-itest
-===
-
-`cf-bridge` integration tests.
