+++ test/cf/multiple-apps/.npmrc
