+++ test/cf/multiple-apps/README.md
@@ -1,0 +1,4 @@
+abacus-cf-multiple-apps-itest
+===
+
+`cf-bridge` and `cf-renewer` multiple applications integration tests.
