+++ test/cf/multiple-apps/src/index.js
@@ -1,0 +1,4 @@
+'use strict';
+
+// There is nothing here - see test/*test.js for the integration and
+// performance tests.
