+++ test/cf/multiple-apps/src/test/test.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const commander = require('commander');
 const cp = require('child_process');
+const commander = require('commander');
 const jwt = require('jsonwebtoken');
 const util = require('util');
-const moment = require('abacus-moment');
 
 const _ = require('underscore');
 const clone = _.clone;
@@ -13,16 +12,17 @@ const dbclient = require('abacus-dbclient');
 const express = require('abacus-express');
 const request = require('abacus-request');
 const router = require('abacus-router');
+const moment = require('abacus-moment');
 
 // Setup the debug log
 const debug =
-  require('abacus-debug')('abacus-cf-bridge-itest');
+  require('abacus-debug')('abacus-cf-multiple-apps-itest');
 const responseDebug =
-  require('abacus-debug')('abacus-cf-bridge-itest-response');
+  require('abacus-debug')('abacus-cf-multiple-apps-itest-response');
 const resultDebug =
-  require('abacus-debug')('abacus-cf-bridge-itest-result');
+  require('abacus-debug')('abacus-cf-multiple-apps-itest-result');
 const oAuthDebug =
-  require('abacus-debug')('abacus-cf-bridge-itest-oauth');
+  require('abacus-debug')('abacus-cf-multiple-apps-itest-oauth');
 
 // Module directory
 const moduleDir = (module) => {
@@ -36,22 +36,6 @@ const timeWindows = {
   'hour'   : 2,
   'day'    : 3,
   'month'  : 4
-};
-
-// Checks if the difference between start and end time fall within a window
-const isWithinWindow = (start, end, timeWindow) => {
-  // [Second, Minute, Hour, Day, Month, Year]
-  const timescale = [1, 100, 10000, 1000000, 100000000, 10000000000];
-  // Converts a millisecond number to a format a number that is YYYYMMDDHHmmSS
-  const dateUTCNumbify = (t) => {
-    const d = moment(t).toDate();
-    return d.getUTCFullYear() * timescale[5] + d.getUTCMonth() * timescale[4]
-      + d.getUTCDate() * timescale[3] + d.getUTCHours() * timescale[2]
-      + d.getUTCMinutes() * timescale[1] + d.getUTCSeconds();
-  };
-
-  return Math.floor(dateUTCNumbify(end) / timescale[timeWindow]) -
-    Math.floor(dateUTCNumbify(start) / timescale[timeWindow]) === 0;
 };
 
 // Parse command line options
@@ -80,7 +64,7 @@ const resourceToken = {
   },
   payload: {
     jti: '254abca5-1c25-40c5-99d7-2cc641791517',
-    sub: 'abacus-cf-bridge',
+    sub: 'abacus-cf-renewer',
     authorities: [
       'abacus.usage.linux-container.write',
       'abacus.usage.linux-container.read'
@@ -89,9 +73,9 @@ const resourceToken = {
       'abacus.usage.linux-container.read',
       'abacus.usage.linux-container.write'
     ],
-    client_id: 'abacus-cf-bridge',
-    cid: 'abacus-cf-bridge',
-    azp: 'abacus-cf-bridge',
+    client_id: 'abacus-cf-renewer',
+    cid: 'abacus-cf-renewer',
+    azp: 'abacus-cf-renewer',
     grant_type: 'client_credentials',
     rev_sig: '2cf89595',
     iat: 1456147679,
@@ -99,11 +83,11 @@ const resourceToken = {
     iss: 'https://localhost:1234/oauth/token',
     zid: 'uaa',
     aud: [
-      'abacus-cf-bridge',
+      'abacus-cf-renewer',
       'abacus.usage.linux-container'
     ]
   },
-  signature: 'irxoV230hkDJenXoTSHQFfqzoUl353lS2URo1fJm21Y'
+  signature: '7BVRprw-yySpW7lSkM8KPZoUIw2w61bs87l0YXqUT8E'
 };
 const systemToken = {
   header: {
@@ -111,7 +95,7 @@ const systemToken = {
   },
   payload: {
     jti: '254abca5-1c25-40c5-99d7-2cc641791517',
-    sub: 'abacus-cf-bridge',
+    sub: 'abacus-cf-renewer',
     authorities: [
       'abacus.usage.write',
       'abacus.usage.read'
@@ -120,9 +104,9 @@ const systemToken = {
       'abacus.usage.write',
       'abacus.usage.read'
     ],
-    client_id: 'abacus-cf-bridge',
-    cid: 'abacus-cf-bridge',
-    azp: 'abacus-cf-bridge',
+    client_id: 'abacus-cf-renewer',
+    cid: 'abacus-cf-renewer',
+    azp: 'abacus-cf-renewer',
     grant_type: 'client_credentials',
     rev_sig: '2cf89595',
     iat: 1456147679,
@@ -130,11 +114,11 @@ const systemToken = {
     iss: 'https://localhost:1234/oauth/token',
     zid: 'uaa',
     aud: [
-      'abacus-cf-bridge',
+      'abacus-cf-renewer',
       'abacus.usage'
     ]
   },
-  signature: 'OVNTKTvu-yHI6QXmYxtPeJZofNddX36Mx1q4PDWuYQE'
+  signature: '1J3_hBJBUgwRO9fzg25sdDYj6DqCVWCNB3veyIBsklM'
 };
 const signedResourceToken = jwt.sign(resourceToken.payload, tokenSecret, {
   expiresIn: 43200
@@ -143,32 +127,32 @@ const signedSystemToken = jwt.sign(systemToken.payload, tokenSecret, {
   expiresIn: 43200
 });
 
-const twentySecondsInMilliseconds = 20 * 1000;
+const lastMonthInMilliseconds = moment().subtract(1, 'months').valueOf();
 
-const test = (secured) => {
-  const submittime = moment.now();
-
+describe('abacus-cf multiple-apps-test with oAuth', () => {
   let server;
   let serverPort;
   let appUsageEvents;
+
+  let noUsageExpected;
   let expectedConsuming;
 
+  const start = (module) => {
+    debug('Starting %s in directory %s', module, moduleDir(module));
+    const c = cp.spawn('npm', ['run', 'start'], {
+      cwd: moduleDir(module),
+      env: clone(process.env)
+    });
+
+    // Add listeners to stdout, stderr and exit message and forward the
+    // messages to debug logs
+    c.stdout.on('data', (data) => process.stdout.write(data));
+    c.stderr.on('data', (data) => process.stderr.write(data));
+    c.on('exit', (code) => debug('Module %s started with code %d',
+      module, code));
+  };
+
   beforeEach((done) => {
-    const start = (module) => {
-      debug('Starting %s in directory %s', module, moduleDir(module));
-      const c = cp.spawn('npm', ['run', 'start'], {
-        cwd: moduleDir(module),
-        env: clone(process.env)
-      });
-
-      // Add listeners to stdout, stderr and exit message and forward the
-      // messages to debug logs
-      c.stdout.on('data', (data) => process.stdout.write(data));
-      c.stderr.on('data', (data) => process.stderr.write(data));
-      c.on('exit', (code) => debug('Module %s started with code %d',
-        module, code));
-    };
-
     const app = express();
     const routes = router();
     routes.get('/v2/app_usage_events', (request, response) => {
@@ -217,28 +201,31 @@ const test = (secured) => {
     serverPort = server.address().port;
     debug('Test resources server listening on port %d', serverPort);
 
-    // Enable/disable the oAuth token authorization
-    process.env.SECURED = secured ? 'true' : 'false';
-    debug('Set SECURED = %s', process.env.SECURED);
+    // Enable the oAuth token authorization
+    process.env.SECURED = 'true';
 
-    // Set environment variables
+    // Secure environment variables
     process.env.API = 'http://localhost:' + serverPort;
     process.env.AUTH_SERVER = 'http://localhost:' + serverPort;
-    process.env.CF_CLIENT_ID = 'abacus-cf-bridge';
+    process.env.CF_CLIENT_ID = 'abacus-cf-renewer';
     process.env.CF_CLIENT_SECRET = 'secret';
     process.env.CLIENT_ID = 'abacus-linux-container';
     process.env.CLIENT_SECRET = 'secret';
+    process.env.ABACUS_CLIENT_ID = 'abacus-cf-renewer';
+    process.env.ABACUS_CLIENT_SECRET = 'secret';
     process.env.JWTKEY = tokenSecret;
     process.env.JWTALGO = tokenAlgorithm;
 
-    // Set slack window to 5 days
-    process.env.SLACK = '5D';
+    // Change slack window to be able to submit usage for last month
+    process.env.SLACK = '32D';
 
-    // Disable wait for correct app-event ordering
-    process.env.GUID_MIN_AGE = twentySecondsInMilliseconds;
+    // Trigger renewer every 2 seconds
+    process.env.RETRY_INTERVAL = 2000;
+
+    noUsageExpected = false;
 
     // Start all Abacus services
-    const services = () => {
+    const startServices = () => {
       start('abacus-eureka-plugin');
       start('abacus-provisioning-plugin');
       start('abacus-account-plugin');
@@ -255,17 +242,17 @@ const test = (secured) => {
     // Start local database server
     if (!process.env.DB) {
       start('abacus-pouchserver');
-      services();
+      startServices();
     }
     else
-      // Delete test dbs on the configured db server
+    // Delete test dbs on the configured db server
       dbclient.drop(process.env.DB, /^abacus-/, () => {
-        services();
+        startServices();
       });
   });
 
   afterEach((done) => {
-    let counter = 10;
+    let counter = 11;
     const finishCb = (module, code) => {
       counter--;
       debug('Module %s exited with code %d. Left %d modules',
@@ -288,6 +275,7 @@ const test = (secured) => {
       c.on('exit', (code) => cb(module, code));
     };
 
+    stop('abacus-cf-renewer', finishCb);
     stop('abacus-cf-bridge', finishCb);
     stop('abacus-usage-reporting', finishCb);
     stop('abacus-usage-aggregator', finishCb);
@@ -311,20 +299,48 @@ const test = (secured) => {
     delete process.env.JWTKEY;
     delete process.env.JWTALGO;
     delete process.env.SLACK;
-    delete process.env.GUID_MIN_AGE;
+    delete process.env.RETRY_INTERVAL;
   });
 
-  const checkAllTimeWindows = (usage, reporttime, level) => {
-    for (const windowType in timeWindows)
-      if(isWithinWindow(submittime, reporttime, timeWindows[windowType])) {
-        const windowUsage = usage.windows[timeWindows[windowType]];
-        if(level !== 'resource')
-          expect(windowUsage[0].quantity.consuming).to.equal(expectedConsuming);
-        expect(windowUsage[0].charge).to.be.above(0);
-      }
+  const checkMonthWindow = (windowName, monthIndex, usage, level) => {
+    const monthWindow = usage.windows[timeWindows.month][monthIndex];
+    expect(monthWindow).to.not.equal(undefined);
+
+    if (level !== 'resource') {
+      expect(monthWindow).to.contain.all.keys('quantity', 'charge');
+      debug('%s window; Expected: consuming=%d, charge>0; ' +
+        'Actual: consuming=%d, charge=%d; Month window: %o',
+        windowName, expectedConsuming, monthWindow.quantity.consuming,
+        monthWindow.charge, monthWindow);
+      expect(monthWindow.quantity.consuming).to.equal(expectedConsuming);
+    }
+    else
+      debug('%s window; Expected:  charge>0; ' +
+        'Actual: charge=%o; Month window: %o',
+        windowName, monthWindow.charge, monthWindow);
+
+    expect(monthWindow).to.contain.all.keys('charge');
+    expect(monthWindow.charge).not.to.equal(undefined);
+    expect(monthWindow.charge).to.be.above(0);
   };
 
-  const checkReport = (cb) => {
+  const checkThisMonth = (windowName, usage, level) => {
+    if (noUsageExpected) {
+      const monthWindow = usage.windows[timeWindows.month][0];
+      expect(monthWindow).to.not.equal(undefined);
+      debug('%s window; Expected: no usage; Actual: %j',
+        windowName, monthWindow);
+      expect(monthWindow).to.equal(null);
+    }
+    else
+      checkMonthWindow(windowName, 0, usage, level);
+  };
+
+  const checkLastMonth = (windowName, usage, level) => {
+    checkMonthWindow(windowName, 1, usage, level);
+  };
+
+  const checkReport = (cb, checkFn) => {
     request.get('http://localhost:9088/v1/metering/organizations' +
       '/:organization_id/aggregated/usage', {
         organization_id: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
@@ -340,16 +356,15 @@ const test = (secured) => {
           const resources = response.body.resources;
           expect(resources.length).to.equal(1);
           expect(response.body.spaces.length).to.equal(1);
-          const reporttime = moment.now();
 
           expect(resources[0]).to.contain.all.keys(
             'plans', 'aggregated_usage');
 
           const planUsage = resources[0].plans[0].aggregated_usage[0];
-          checkAllTimeWindows(planUsage, reporttime);
+          checkFn('Plans aggregated usage', planUsage);
 
           const aggregatedUsage = resources[0].aggregated_usage[0];
-          checkAllTimeWindows(aggregatedUsage, reporttime, 'resource');
+          checkFn('Aggregated usage', aggregatedUsage, 'resource');
 
           resultDebug('All usage report checks are successful for: %s',
             JSON.stringify(response.body, null, 2));
@@ -366,7 +381,7 @@ const test = (secured) => {
       });
   };
 
-  const poll = (fn, done, timeout = 1000, interval = 100) => {
+  const poll = (fn, checkFn, done, timeout = 1000, interval = 100) => {
     const startTimestamp = moment.now();
 
     const doneCallback = (err) => {
@@ -383,16 +398,15 @@ const test = (secured) => {
       else
         setTimeout(() => {
           debug('Calling %s after >= %d ms...', fn.name, interval);
-          fn(doneCallback);
+          fn(doneCallback, checkFn);
         }, interval);
     };
 
     debug('Calling %s for the first time...', fn.name);
-    fn(doneCallback);
+    fn(doneCallback, checkFn);
   };
 
-  const waitForStartAndPoll = (component, port, done) => {
-    // Wait for bridge to start
+  const waitForStartAndPoll = (component, port, checkFn, timeout, done) => {
     let startWaitTime = moment.now();
     request.waitFor('http://localhost::p/v1/cf/:component',
       { component: component, p: port },
@@ -403,31 +417,33 @@ const test = (secured) => {
         // Check report
         request.get(uri, {
           headers: {
-            authorization: secured ? 'bearer ' + signedSystemToken : ''
+            authorization: 'bearer ' + signedSystemToken
           }
         }, (err, response) => {
           expect(err).to.equal(undefined);
           expect(response.statusCode).to.equal(200);
 
-          poll(checkReport, (error) => {
+          const t = timeout - (moment.now() - startWaitTime);
+          debug('Time left for executing test: %d ms', t);
+          poll(checkReport, checkFn, (error) => {
             done(error);
-          }, totalTimeout - (moment.now() - startWaitTime), 1000);
+          }, t, 1000);
         });
       }
     );
   };
 
-  context('with a single app', () => {
+  context('with multiple apps', () => {
 
     context('start, stop, start, scale out', () => {
       beforeEach(() => {
         appUsageEvents = [
+          // app1 start
           {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds).toISOString()
+              created_at: new moment(lastMonthInMilliseconds).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -452,12 +468,12 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app1 buildpack set
           {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 1).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 1).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
@@ -482,19 +498,49 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app2 start
+          {
+            metadata: {
+              guid: 'b557f9e6-19f6-4263-9ffe-be39feccd577',
+              url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
+              created_at: moment(lastMonthInMilliseconds + 2).toISOString()
+            },
+            entity: {
+              state: 'STARTED',
+              previous_state: 'STOPPED',
+              memory_in_mb_per_instance: 512,
+              previous_memory_in_mb_per_instance: 512,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: null,
+              buildpack_name: null,
+              package_state: 'PENDING',
+              previous_package_state: 'PENDING',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 stop
           {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 2).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 3).toISOString()
             },
             entity: {
               state: 'STOPPED',
               previous_state: 'STARTED',
               memory_in_mb_per_instance: 512,
               previous_memory_in_mb_per_instance: 512,
-              instance_count: 2,
+              instance_count: 1,
               previous_instance_count: 1,
               app_guid: '35c4ff2f',
               app_name: 'app',
@@ -512,12 +558,42 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app2 buildpack set
+          {
+            metadata: {
+              guid: '1f2336af-1866-4d2b-8845-0efb14c1a389',
+              url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
+              created_at: moment(lastMonthInMilliseconds + 4).toISOString()
+            },
+            entity: {
+              state: 'BUILDPACK_SET',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 512,
+              previous_memory_in_mb_per_instance: 512,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 start #2
           {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 3).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 5).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -542,12 +618,72 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app2 stop
+          {
+            metadata: {
+              guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
+              url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
+              created_at: moment(lastMonthInMilliseconds + 6).toISOString()
+            },
+            entity: {
+              state: 'STOPPED',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 512,
+              previous_memory_in_mb_per_instance: 512,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app2 scale
+          {
+            metadata: {
+              guid: '358ea444-943d-4a6e-9928-786a5bb93dfb',
+              url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
+              created_at: moment(lastMonthInMilliseconds + 7).toISOString()
+            },
+            entity: {
+              state: 'STARTED',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 512,
+              instance_count: 2,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 buildpack set #2
           {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 4).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 8).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
@@ -572,12 +708,42 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app1 stop
           {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 5).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 9).toISOString()
+            },
+            entity: {
+              state: 'STOPPED',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 512,
+              previous_memory_in_mb_per_instance: 512,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '35c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 scale
+          {
+            metadata: {
+              guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
+              url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
+              created_at: moment(lastMonthInMilliseconds + 10).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -604,29 +770,45 @@ const test = (secured) => {
           }
         ];
 
-        // first start: 0.5 GB
-        // second start: 0.25 GB
+        // app1 first start: 0.5 GB
+        // app1 second start: 0.25 GB
+        // app1 scale out: 2 x 1 GB = 2GB
+        // app2 first start: 0.5 GB
+        // app2 second start: 0.25 GB
+        // app2 scale out: 2 x 1 GB = 2GB
+        //
         // buildpack_set events are ignored
-        // scale out: 2 x 1 GB = 2GB
-        expectedConsuming = 2;
+        expectedConsuming = 4;
       });
 
-      it('submits usage and gets expected report back', function(done) {
+      it('submits runtime usage to usage collector', function(done) {
         this.timeout(totalTimeout + 2000);
 
-        waitForStartAndPoll('bridge', 9500, done);
+        const startTestTime = moment.now();
+        waitForStartAndPoll('bridge', 9500, checkLastMonth, totalTimeout,
+          (error) => {
+            if (error) {
+              done(error);
+              return;
+            }
+            process.env.slack = '5D';
+            start('abacus-cf-renewer');
+            waitForStartAndPoll('renewer', 9501, checkThisMonth,
+              totalTimeout - (moment.now() - startTestTime), done);
+          }
+        );
       });
     });
 
     context('start, scale out, stop', () => {
       beforeEach(() => {
         appUsageEvents = [
+          // app1 start
           {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds).toISOString()
+              created_at: moment(lastMonthInMilliseconds).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -651,12 +833,12 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app1 buildpack set
           {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 1).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 1).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
@@ -681,12 +863,73 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app2 start
+          {
+            metadata: {
+              guid: 'b557f9e6-19f6-4263-9ffe-be39feccd577',
+              url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
+              created_at: moment(lastMonthInMilliseconds + 2).toISOString()
+            },
+            entity: {
+              state: 'STARTED',
+              previous_state: 'STOPPED',
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 1024,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: null,
+              buildpack_name: null,
+              package_state: 'PENDING',
+              previous_package_state: 'PENDING',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app2 buildpack set
+          {
+            metadata: {
+              guid: '1f2336af-1866-4d2b-8845-0efb14c1a389',
+              url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
+              created_at: moment(lastMonthInMilliseconds -
+                + 3).toISOString()
+            },
+            entity: {
+              state: 'BUILDPACK_SET',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 1024,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 scale out
           {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 2).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 4).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -711,12 +954,42 @@ const test = (secured) => {
               task_guid: null
             }
           },
+          // app2 stop
+          {
+            metadata: {
+              guid: '458ea444-943d-4a6e-9928-786a5bb93dfb',
+              url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
+              created_at: moment(lastMonthInMilliseconds + 5).toISOString()
+            },
+            entity: {
+              state: 'STOPPED',
+              previous_state: 'STARTED',
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 1024,
+              instance_count: 1,
+              previous_instance_count: 1,
+              app_guid: '45c4ff2f',
+              app_name: 'app',
+              space_guid: 'a7e44fcd-25bf-4023-8a87-03fba4882995',
+              space_name: 'abacus',
+              org_guid: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
+              buildpack_guid: '30429b05-745e-4474-a39f-267afa365d69',
+              buildpack_name: 'staticfile_buildpack',
+              package_state: 'STAGED',
+              previous_package_state: 'STAGED',
+              parent_app_guid: null,
+              parent_app_name: null,
+              process_type: 'web',
+              task_name: null,
+              task_guid: null
+            }
+          },
+          // app1 stop
           {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(submittime -
-                twentySecondsInMilliseconds + 3).toISOString()
+              created_at: moment(lastMonthInMilliseconds + 6).toISOString()
             },
             entity: {
               state: 'STOPPED',
@@ -743,23 +1016,33 @@ const test = (secured) => {
           }
         ];
 
-        // first start: 1 GB
-        // scale: 2x2GB = 4GB
+        // app1 start: 1 GB
+        // app1 scale: 2x2GB = 4GB
+        // app1 stop: 0GB
+        // app2 start: 1 GB
+        // app2 stop: 0GB
+        //
         // buildpack_set events are ignored
-        // stop: 0GB
         expectedConsuming = 0;
+        noUsageExpected = true;
       });
 
-      it('submits usage and gets expected report back', function(done) {
-        this.timeout(totalTimeout + 2000);
+      it('submits runtime usage to usage collector', function(done) {
+        this.timeout(totalTimeout);
 
-        waitForStartAndPoll('bridge', 9500, done);
+        const startTestTime = moment.now();
+        waitForStartAndPoll('bridge', 9500, checkLastMonth, totalTimeout,
+          (error) => {
+            if (error) {
+              done(error);
+              return;
+            }
+            start('abacus-cf-renewer');
+            waitForStartAndPoll('renewer', 9501, checkThisMonth,
+              totalTimeout - (moment.now() - startTestTime), done);
+          }
+        );
       });
     });
-
   });
-};
-
-describe('abacus-cf-bridge single-app-test without oAuth', () => test(false));
-
-describe('abacus-cf-bridge single-app-test with oAuth', () => test(true));
+});
