+++ test/cf/purge/.gitignore
@@ -1,0 +1,2 @@
+lib/**
+node_modules/**
