+++ test/cf/purge/.npmrc
@@ -1,0 +1,1 @@
+optional=false
