+++ test/cf/purge/README.md
@@ -1,0 +1,4 @@
+abacus-cf-purge-itest
+===
+
+`cf-bridge` purge integration tests.
