+++ test/cf/purge/package.json
@@ -1,0 +1,88 @@
+{
+  "name": "abacus-cf-purge-itest",
+  "description": "CF Bridge purge integration tests",
+  "license": "Apache-2.0",
+  "version": "0.0.6-dev.8",
+  "private": true,
+  "homepage": "https://github.com/cloudfoundry-incubator/cf-abacus/test/cf/purge",
+  "bugs": {
+    "url": "https://github.com/cloudfoundry-incubator/cf-abacus/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "http://github.com/cloudfoundry-incubator/cf-abacus.git"
+  },
+  "keywords": [
+    "cf",
+    "abacus"
+  ],
+  "files": [
+    ".npmrc",
+    "lib/",
+    "src/"
+  ],
+  "main": "lib/index.js",
+  "scripts": {
+    "itest": "mocha --command",
+    "babel": "babel",
+    "test": "eslint",
+    "lint": "eslint",
+    "pub": "publish"
+  },
+  "dependencies": {
+    "abacus-mocha": "file:../../../tools/mocha",
+    "abacus-metering-config": "file:../../../lib/config/metering",
+    "abacus-pricing-config": "file:../../../lib/config/pricing",
+    "abacus-usage-schemas": "file:../../../lib/config/schemas",
+    "abacus-usage-accumulator": "file:../../../lib/aggregation/accumulator",
+    "abacus-usage-aggregator": "file:../../../lib/aggregation/aggregator",
+    "abacus-usage-reporting": "file:../../../lib/aggregation/reporting",
+    "abacus-usage-collector": "file:../../../lib/metering/collector",
+    "abacus-usage-meter": "file:../../../lib/metering/meter",
+    "abacus-account-plugin": "file:../../../lib/plugins/account",
+    "abacus-provisioning-plugin": "file:../../../lib/plugins/provisioning",
+    "abacus-eureka-plugin": "file:../../../lib/plugins/eureka",
+    "abacus-cf-bridge": "file:../../../lib/cf/bridge",
+    "abacus-batch": "file:../../../lib/utils/batch",
+    "abacus-breaker": "file:../../../lib/utils/breaker",
+    "abacus-oauth": "file:../../../lib/utils/oauth",
+    "abacus-cluster": "file:../../../lib/utils/cluster",
+    "abacus-dbclient": "file:../../../lib/utils/dbclient",
+    "abacus-pouchserver": "file:../../../lib/utils/pouchserver",
+    "abacus-debug": "file:../../../lib/utils/debug",
+    "abacus-eval": "file:../../../lib/utils/eval",
+    "abacus-events": "file:../../../lib/utils/events",
+    "abacus-express": "file:../../../lib/utils/express",
+    "abacus-hystrix": "file:../../../lib/utils/hystrix",
+    "abacus-launcher": "file:../../../lib/utils/launcher",
+    "abacus-lock": "file:../../../lib/utils/lock",
+    "abacus-partition": "file:../../../lib/utils/partition",
+    "abacus-perf": "file:../../../lib/utils/perf",
+    "abacus-request": "file:../../../lib/utils/request",
+    "abacus-retry": "file:../../../lib/utils/retry",
+    "abacus-router": "file:../../../lib/utils/router",
+    "abacus-schema": "file:../../../lib/utils/schema",
+    "abacus-seqid": "file:../../../lib/utils/seqid",
+    "abacus-throttle": "file:../../../lib/utils/throttle",
+    "abacus-transform": "file:../../../lib/utils/transform",
+    "abacus-urienv": "file:../../../lib/utils/urienv",
+    "abacus-vcapenv": "file:../../../lib/utils/vcapenv",
+    "abacus-webapp": "file:../../../lib/utils/webapp",
+    "abacus-yieldable": "file:../../../lib/utils/yieldable",
+    "babel-preset-es2015": "^6.6.0",
+    "abacus-doc": "file:../../../doc",
+    "abacus-etc": "file:../../../etc",
+    "underscore": "^1.8.3",
+    "commander": "2.8.1",
+    "jsonwebtoken": "^5.0.5"
+  },
+  "devDependencies": {
+    "abacus-babel": "file:../../../tools/babel",
+    "abacus-eslint": "file:../../../tools/eslint",
+    "abacus-publish": "file:../../../tools/publish"
+  },
+  "engines": {
+    "node": ">=5.11.1",
+    "npm": ">=3.8.6"
+  }
+}
