+++ test/cf/single-app/README.md
@@ -1,0 +1,4 @@
+abacus-cf-single-app-itest
+===
+
+`cf-bridge` and `cf-renewer` single application integration tests.
