+++ test/cf/single-app/src/test/bridge-test.js
@@ -1,9 +1,10 @@
 'use strict';
 
-const cp = require('child_process');
 const commander = require('commander');
+const cp = require('child_process');
 const jwt = require('jsonwebtoken');
 const util = require('util');
+const moment = require('abacus-moment');
 
 const _ = require('underscore');
 const clone = _.clone;
@@ -12,17 +13,16 @@ const dbclient = require('abacus-dbclient');
 const express = require('abacus-express');
 const request = require('abacus-request');
 const router = require('abacus-router');
-const moment = require('abacus-moment');
 
 // Setup the debug log
 const debug =
-  require('abacus-debug')('abacus-cf-renewer-itest');
+  require('abacus-debug')('abacus-cf-single-app-itest');
 const responseDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-response');
+  require('abacus-debug')('abacus-cf-single-app-itest-response');
 const resultDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-result');
+  require('abacus-debug')('abacus-cf-single-app-itest-result');
 const oAuthDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-oauth');
+  require('abacus-debug')('abacus-cf-single-app-itest-oauth');
 
 // Module directory
 const moduleDir = (module) => {
@@ -36,6 +36,22 @@ const timeWindows = {
   'hour'   : 2,
   'day'    : 3,
   'month'  : 4
+};
+
+// Checks if the difference between start and end time fall within a window
+const isWithinWindow = (start, end, timeWindow) => {
+  // [Second, Minute, Hour, Day, Month, Year]
+  const timescale = [1, 100, 10000, 1000000, 100000000, 10000000000];
+  // Converts a millisecond number to a format a number that is YYYYMMDDHHmmSS
+  const dateUTCNumbify = (t) => {
+    const d = moment(t).toDate();
+    return d.getUTCFullYear() * timescale[5] + d.getUTCMonth() * timescale[4]
+      + d.getUTCDate() * timescale[3] + d.getUTCHours() * timescale[2]
+      + d.getUTCMinutes() * timescale[1] + d.getUTCSeconds();
+  };
+
+  return Math.floor(dateUTCNumbify(end) / timescale[timeWindow]) -
+    Math.floor(dateUTCNumbify(start) / timescale[timeWindow]) === 0;
 };
 
 // Parse command line options
@@ -64,7 +80,7 @@ const resourceToken = {
   },
   payload: {
     jti: '254abca5-1c25-40c5-99d7-2cc641791517',
-    sub: 'abacus-cf-renewer',
+    sub: 'abacus-cf-bridge',
     authorities: [
       'abacus.usage.linux-container.write',
       'abacus.usage.linux-container.read'
@@ -73,9 +89,9 @@ const resourceToken = {
       'abacus.usage.linux-container.read',
       'abacus.usage.linux-container.write'
     ],
-    client_id: 'abacus-cf-renewer',
-    cid: 'abacus-cf-renewer',
-    azp: 'abacus-cf-renewer',
+    client_id: 'abacus-cf-bridge',
+    cid: 'abacus-cf-bridge',
+    azp: 'abacus-cf-bridge',
     grant_type: 'client_credentials',
     rev_sig: '2cf89595',
     iat: 1456147679,
@@ -83,11 +99,11 @@ const resourceToken = {
     iss: 'https://localhost:1234/oauth/token',
     zid: 'uaa',
     aud: [
-      'abacus-cf-renewer',
+      'abacus-cf-bridge',
       'abacus.usage.linux-container'
     ]
   },
-  signature: '7BVRprw-yySpW7lSkM8KPZoUIw2w61bs87l0YXqUT8E'
+  signature: 'irxoV230hkDJenXoTSHQFfqzoUl353lS2URo1fJm21Y'
 };
 const systemToken = {
   header: {
@@ -95,7 +111,7 @@ const systemToken = {
   },
   payload: {
     jti: '254abca5-1c25-40c5-99d7-2cc641791517',
-    sub: 'abacus-cf-renewer',
+    sub: 'abacus-cf-bridge',
     authorities: [
       'abacus.usage.write',
       'abacus.usage.read'
@@ -104,9 +120,9 @@ const systemToken = {
       'abacus.usage.write',
       'abacus.usage.read'
     ],
-    client_id: 'abacus-cf-renewer',
-    cid: 'abacus-cf-renewer',
-    azp: 'abacus-cf-renewer',
+    client_id: 'abacus-cf-bridge',
+    cid: 'abacus-cf-bridge',
+    azp: 'abacus-cf-bridge',
     grant_type: 'client_credentials',
     rev_sig: '2cf89595',
     iat: 1456147679,
@@ -114,11 +130,11 @@ const systemToken = {
     iss: 'https://localhost:1234/oauth/token',
     zid: 'uaa',
     aud: [
-      'abacus-cf-renewer',
+      'abacus-cf-bridge',
       'abacus.usage'
     ]
   },
-  signature: '1J3_hBJBUgwRO9fzg25sdDYj6DqCVWCNB3veyIBsklM'
+  signature: 'OVNTKTvu-yHI6QXmYxtPeJZofNddX36Mx1q4PDWuYQE'
 };
 const signedResourceToken = jwt.sign(resourceToken.payload, tokenSecret, {
   expiresIn: 43200
@@ -127,33 +143,32 @@ const signedSystemToken = jwt.sign(systemToken.payload, tokenSecret, {
   expiresIn: 43200
 });
 
-const lastMonthInMilliseconds = moment().utc().subtract(1, 'months').valueOf();
+const twentySecondsInMilliseconds = 20 * 1000;
 
-const test = (secured) => {
+describe('abacus-cf-bridge single-app-test without oAuth', () => {
+  const submittime = moment.now();
+
   let server;
   let serverPort;
-
   let appUsageEvents;
-
-  let noUsageExpected;
   let expectedConsuming;
 
-  const start = (module) => {
-    debug('Starting %s in directory %s', module, moduleDir(module));
-    const c = cp.spawn('npm', ['run', 'start'], {
-      cwd: moduleDir(module),
-      env: clone(process.env)
-    });
-
-    // Add listeners to stdout, stderr and exit message and forward the
-    // messages to debug logs
-    c.stdout.on('data', (data) => process.stdout.write(data));
-    c.stderr.on('data', (data) => process.stderr.write(data));
-    c.on('exit', (code) => debug('Module %s started with code %d',
-      module, code));
-  };
-
   beforeEach((done) => {
+    const start = (module) => {
+      debug('Starting %s in directory %s', module, moduleDir(module));
+      const c = cp.spawn('npm', ['run', 'start'], {
+        cwd: moduleDir(module),
+        env: clone(process.env)
+      });
+
+      // Add listeners to stdout, stderr and exit message and forward the
+      // messages to debug logs
+      c.stdout.on('data', (data) => process.stdout.write(data));
+      c.stderr.on('data', (data) => process.stderr.write(data));
+      c.on('exit', (code) => debug('Module %s started with code %d',
+        module, code));
+    };
+
     const app = express();
     const routes = router();
     routes.get('/v2/app_usage_events', (request, response) => {
@@ -169,7 +184,6 @@ const test = (secured) => {
         return;
       }
 
-      responseDebug('Returning events %j', appUsageEvents);
       response.status(200).send({
         total_results: appUsageEvents.length,
         total_pages: 1,
@@ -203,32 +217,24 @@ const test = (secured) => {
     serverPort = server.address().port;
     debug('Test resources server listening on port %d', serverPort);
 
-    // Enable/disable the oAuth token authorization
-    process.env.SECURED = secured ? 'true' : 'false';
-    debug('Set SECURED = %s', process.env.SECURED);
-
     // Set environment variables
     process.env.API = 'http://localhost:' + serverPort;
     process.env.AUTH_SERVER = 'http://localhost:' + serverPort;
-    process.env.CF_CLIENT_ID = 'abacus-cf-renewer';
+    process.env.CF_CLIENT_ID = 'abacus-cf-bridge';
     process.env.CF_CLIENT_SECRET = 'secret';
     process.env.CLIENT_ID = 'abacus-linux-container';
     process.env.CLIENT_SECRET = 'secret';
-    process.env.ABACUS_CLIENT_ID = 'abacus-cf-renewer';
-    process.env.ABACUS_CLIENT_SECRET = 'secret';
     process.env.JWTKEY = tokenSecret;
     process.env.JWTALGO = tokenAlgorithm;
 
-    // Change slack window to be able to submit usage for last month
-    process.env.SLACK = '32D';
+    // Set slack window to 5 days
+    process.env.SLACK = '5D';
 
-    // Trigger renewer every 2 seconds
-    process.env.RETRY_INTERVAL = 2000;
-
-    noUsageExpected = false;
+    // Disable wait for correct app-event ordering
+    process.env.GUID_MIN_AGE = twentySecondsInMilliseconds;
 
     // Start all Abacus services
-    const startServices = () => {
+    const services = () => {
       start('abacus-eureka-plugin');
       start('abacus-provisioning-plugin');
       start('abacus-account-plugin');
@@ -245,17 +251,17 @@ const test = (secured) => {
     // Start local database server
     if (!process.env.DB) {
       start('abacus-pouchserver');
-      startServices();
+      services();
     }
     else
       // Delete test dbs on the configured db server
       dbclient.drop(process.env.DB, /^abacus-/, () => {
-        startServices();
+        services();
       });
   });
 
   afterEach((done) => {
-    let counter = 11;
+    let counter = 10;
     const finishCb = (module, code) => {
       counter--;
       debug('Module %s exited with code %d. Left %d modules',
@@ -278,7 +284,6 @@ const test = (secured) => {
       c.on('exit', (code) => cb(module, code));
     };
 
-    stop('abacus-cf-renewer', finishCb);
     stop('abacus-cf-bridge', finishCb);
     stop('abacus-usage-reporting', finishCb);
     stop('abacus-usage-aggregator', finishCb);
@@ -302,65 +307,20 @@ const test = (secured) => {
     delete process.env.JWTKEY;
     delete process.env.JWTALGO;
     delete process.env.SLACK;
-    delete process.env.RETRY_INTERVAL;
+    delete process.env.GUID_MIN_AGE;
   });
 
-  const checkLastMonthWindow = (windowName, usage, level) => {
-    const windowUsage = usage.windows[timeWindows.month];
-    const previousMonth = windowUsage[1];
-
-    expect(previousMonth).to.not.equal(undefined);
-
-    if (level !== 'resource') {
-      expect(previousMonth).to.contain.all.keys('quantity', 'charge');
-      debug('%s month window; Expected: consuming=%d, charge>0; ' +
-        'Actual: consuming=%d, charge=%d; Month window: %o', windowName,
-        expectedConsuming, previousMonth.quantity.consuming,
-        previousMonth.charge, previousMonth);
-      expect(previousMonth.quantity.consuming).to.equal(expectedConsuming);
-    }
-    else
-      debug('%s month window; Expected: charge>0; ' +
-        'Actual: charge=%o; Month window: %o',
-        windowName, previousMonth.charge, previousMonth);
-
-    expect(previousMonth).to.contain.all.keys('charge');
-    expect(previousMonth.charge).to.not.equal(undefined);
-    expect(previousMonth.charge).to.be.above(0);
+  const checkAllTimeWindows = (usage, reporttime, level) => {
+    for (const windowType in timeWindows)
+      if(isWithinWindow(submittime, reporttime, timeWindows[windowType])) {
+        const windowUsage = usage.windows[timeWindows[windowType]];
+        if(level !== 'resource')
+          expect(windowUsage[0].quantity.consuming).to.equal(expectedConsuming);
+        expect(windowUsage[0].charge).to.be.above(0);
+      }
   };
 
-  const checkCurrentMonthWindow = (windowName, usage, level) => {
-    const windowUsage = usage.windows[timeWindows.month];
-    const currentMonth = windowUsage[0];
-
-    expect(currentMonth).to.not.equal(undefined);
-
-    if (noUsageExpected) {
-      debug('%s window; Expected: no usage; Actual: %j',
-        windowName, currentMonth);
-      expect(currentMonth).to.equal(null);
-      return;
-    }
-
-    if (level !== 'resource') {
-      expect(currentMonth).to.contain.all.keys('quantity', 'charge');
-      debug('%s window; Expected: consuming=%d, charge>0; ' +
-        'Actual: consuming=%d, charge=%d; Month window: %o',
-        windowName, expectedConsuming, currentMonth.quantity.consuming,
-        currentMonth.charge, currentMonth);
-      expect(currentMonth.quantity.consuming).to.equal(expectedConsuming);
-    }
-    else
-      debug('%s window; Expected:  charge>0; ' +
-        'Actual: charge=%o; Month window: %o',
-        windowName, currentMonth.charge, currentMonth);
-
-    expect(currentMonth.charge).not.to.equal(undefined);
-    expect(currentMonth).to.contain.all.keys('charge');
-    expect(currentMonth.charge).to.be.above(0);
-  };
-
-  const checkReport = (cb, checkFn) => {
+  const checkReport = (cb) => {
     request.get('http://localhost:9088/v1/metering/organizations' +
       '/:organization_id/aggregated/usage', {
         organization_id: 'e8139b76-e829-4af3-b332-87316b1c0a6c',
@@ -376,15 +336,16 @@ const test = (secured) => {
           const resources = response.body.resources;
           expect(resources.length).to.equal(1);
           expect(response.body.spaces.length).to.equal(1);
+          const reporttime = moment.now();
 
           expect(resources[0]).to.contain.all.keys(
             'plans', 'aggregated_usage');
 
           const planUsage = resources[0].plans[0].aggregated_usage[0];
-          checkFn('Plans aggregated usage', planUsage);
+          checkAllTimeWindows(planUsage, reporttime);
 
           const aggregatedUsage = resources[0].aggregated_usage[0];
-          checkFn('Aggregated usage', aggregatedUsage, 'resource');
+          checkAllTimeWindows(aggregatedUsage, reporttime, 'resource');
 
           resultDebug('All usage report checks are successful for: %s',
             JSON.stringify(response.body, null, 2));
@@ -401,7 +362,7 @@ const test = (secured) => {
       });
   };
 
-  const poll = (fn, checkFn, done, timeout = 1000, interval = 100) => {
+  const poll = (fn, done, timeout = 1000, interval = 100) => {
     const startTimestamp = moment.now();
 
     const doneCallback = (err) => {
@@ -418,15 +379,16 @@ const test = (secured) => {
       else
         setTimeout(() => {
           debug('Calling %s after >= %d ms...', fn.name, interval);
-          fn(doneCallback, checkFn);
+          fn(doneCallback);
         }, interval);
     };
 
     debug('Calling %s for the first time...', fn.name);
-    fn(doneCallback, checkFn);
+    fn(doneCallback);
   };
 
-  const waitForStartAndPoll = (component, port, checkFn, timeout, done) => {
+  const waitForStartAndPoll = (component, port, done) => {
+    // Wait for bridge to start
     let startWaitTime = moment.now();
     request.waitFor('http://localhost::p/v1/cf/:component',
       { component: component, p: port },
@@ -435,19 +397,13 @@ const test = (secured) => {
         if (err) throw err;
 
         // Check report
-        request.get(uri, {
-          headers: {
-            authorization: secured ? 'bearer ' + signedSystemToken : ''
-          }
-        }, (err, response) => {
+        request.get(uri, {}, (err, response) => {
           expect(err).to.equal(undefined);
           expect(response.statusCode).to.equal(200);
 
-          const t = timeout - (moment.now() - startWaitTime);
-          debug('Time left for executing test: %d ms', t);
-          poll(checkReport, checkFn, (error) => {
+          poll(checkReport, (error) => {
             done(error);
-          }, t, 1000);
+          }, totalTimeout - (moment.now() - startWaitTime), 1000);
         });
       }
     );
@@ -462,7 +418,8 @@ const test = (secured) => {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(lastMonthInMilliseconds).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -491,7 +448,8 @@ const test = (secured) => {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(lastMonthInMilliseconds + 1).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 1).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
@@ -520,7 +478,8 @@ const test = (secured) => {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(lastMonthInMilliseconds + 2).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 2).toISOString()
             },
             entity: {
               state: 'STOPPED',
@@ -549,7 +508,8 @@ const test = (secured) => {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(lastMonthInMilliseconds + 3).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 3).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -578,7 +538,8 @@ const test = (secured) => {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(lastMonthInMilliseconds + 4).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 4).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
@@ -607,7 +568,8 @@ const test = (secured) => {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(lastMonthInMilliseconds + 5).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 5).toISOString()
             },
             entity: {
               state: 'STARTED',
@@ -641,21 +603,10 @@ const test = (secured) => {
         expectedConsuming = 2;
       });
 
-      it('submits runtime usage to usage collector', function(done) {
-        this.timeout(totalTimeout);
+      it('submits usage and gets expected report back', function(done) {
+        this.timeout(totalTimeout + 2000);
 
-        const startTestTime = moment.now();
-        waitForStartAndPoll('bridge', 9500, checkLastMonthWindow, totalTimeout,
-          (error) => {
-            if (error) {
-              done(error);
-              return;
-            }
-            start('abacus-cf-renewer');
-            waitForStartAndPoll('renewer', 9501, checkCurrentMonthWindow,
-              totalTimeout - (moment.now() - startTestTime), done);
-          }
-        );
+        waitForStartAndPoll('bridge', 9500, done);
       });
     });
 
@@ -666,13 +617,14 @@ const test = (secured) => {
             metadata: {
               guid: 'b457f9e6-19f6-4263-9ffe-be39feccd576',
               url: '/v2/app_usage_events/b457f9e6-19f6-4263-9ffe-be39feccd576',
-              created_at: moment(lastMonthInMilliseconds).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds).toISOString()
             },
             entity: {
               state: 'STARTED',
               previous_state: 'STOPPED',
-              memory_in_mb_per_instance: 512,
-              previous_memory_in_mb_per_instance: 512,
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 1024,
               instance_count: 1,
               previous_instance_count: 1,
               app_guid: '35c4ff2f',
@@ -695,13 +647,14 @@ const test = (secured) => {
             metadata: {
               guid: '0f2336af-1866-4d2b-8845-0efb14c1a388',
               url: '/v2/app_usage_events/0f2336af-1866-4d2b-8845-0efb14c1a388',
-              created_at: moment(lastMonthInMilliseconds + 1).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 1).toISOString()
             },
             entity: {
               state: 'BUILDPACK_SET',
               previous_state: 'STARTED',
-              memory_in_mb_per_instance: 512,
-              previous_memory_in_mb_per_instance: 512,
+              memory_in_mb_per_instance: 1024,
+              previous_memory_in_mb_per_instance: 1024,
               instance_count: 1,
               previous_instance_count: 1,
               app_guid: '35c4ff2f',
@@ -724,13 +677,14 @@ const test = (secured) => {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(lastMonthInMilliseconds + 2).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 2).toISOString()
             },
             entity: {
               state: 'STARTED',
               previous_state: 'STARTED',
-              memory_in_mb_per_instance: 1024,
-              previous_memory_in_mb_per_instance: 512,
+              memory_in_mb_per_instance: 2048,
+              previous_memory_in_mb_per_instance: 1024,
               instance_count: 2,
               previous_instance_count: 1,
               app_guid: '35c4ff2f',
@@ -753,13 +707,14 @@ const test = (secured) => {
             metadata: {
               guid: '258ea444-943d-4a6e-9928-786a5bb93dfa',
               url: '/v2/app_usage_events/258ea444-943d-4a6e-9928-786a5bb93dfa',
-              created_at: moment(lastMonthInMilliseconds + 3).toISOString()
+              created_at: moment(submittime -
+                twentySecondsInMilliseconds + 3).toISOString()
             },
             entity: {
               state: 'STOPPED',
               previous_state: 'STARTED',
-              memory_in_mb_per_instance: 1024,
-              previous_memory_in_mb_per_instance: 1024,
+              memory_in_mb_per_instance: 2048,
+              previous_memory_in_mb_per_instance: 2048,
               instance_count: 2,
               previous_instance_count: 2,
               app_guid: '35c4ff2f',
@@ -780,36 +735,19 @@ const test = (secured) => {
           }
         ];
 
-        // start: 0.5 GB
+        // first start: 1 GB
+        // scale: 2x2GB = 4GB
         // buildpack_set events are ignored
-        // scale out: 2 x 1 GB = 2GB
         // stop: 0GB
         expectedConsuming = 0;
-        noUsageExpected = true;
       });
 
-      it('submits runtime usage to usage collector', function(done) {
-        this.timeout(totalTimeout);
+      it('submits usage and gets expected report back', function(done) {
+        this.timeout(totalTimeout + 2000);
 
-        const startTestTime = moment.now();
-        waitForStartAndPoll('bridge', 9500, checkLastMonthWindow, totalTimeout,
-          (error) => {
-            if (error) {
-              done(error);
-              return;
-            }
-            start('abacus-cf-renewer');
-            waitForStartAndPoll('renewer', 9501, checkCurrentMonthWindow,
-              totalTimeout - (moment.now() - startTestTime), done);
-          }
-        );
+        waitForStartAndPoll('bridge', 9500, done);
       });
     });
+
   });
-};
-
-describe('abacus-cf-renewer single-app-test without oAuth',
-  () => test(false));
-
-describe('abacus-cf-renewer single-app-test with oAuth',
-  () => test(true));
+});
