+++ test/cf/slack-window/.gitignore
@@ -1,0 +1,2 @@
+lib/**
+node_modules/**
