+++ test/cf/slack-window/README.md
@@ -1,0 +1,4 @@
+abacus-cf-slack-window-itest
+===
+
+`cf-bridge` and `cf-renewer` slack window integration tests.
