+++ test/cf/slack-window/src/test/renewer-test.js
@@ -16,13 +16,13 @@ const moment = require('abacus-moment');
 
 // Setup the debug log
 const debug =
-  require('abacus-debug')('abacus-cf-renewer-itest');
+  require('abacus-debug')('abacus-cf-renewer-slack-window-itest');
 const responseDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-response');
+  require('abacus-debug')('abacus-cf-renewer-slack-window-itest-response');
 const resultDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-result');
+  require('abacus-debug')('abacus-cf-renewer-slack-window-itest-result');
 const oAuthDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-oauth');
+  require('abacus-debug')('abacus-cf-renewer-slack-window-itest-oauth');
 
 // Module directory
 const moduleDir = (module) => {
