+++ test/cf/timeshift/src/test/renewer-test.js
@@ -19,13 +19,13 @@ const clone = _.clone;
 
 // Setup the debug log
 const debug =
-  require('abacus-debug')('abacus-cf-renewer-itest');
+  require('abacus-debug')('abacus-cf-renewer-time-shift-itest');
 const responseDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-response');
+  require('abacus-debug')('abacus-cf-renewer-time-shift-itest-response');
 const resultDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-result');
+  require('abacus-debug')('abacus-cf-renewer-time-shift-itest-result');
 const oAuthDebug =
-  require('abacus-debug')('abacus-cf-renewer-itest-oauth');
+  require('abacus-debug')('abacus-cf-renewer-time-shift-itest-oauth');
 
 // Parse command line options
 const argv = clone(process.argv);
